Extract 'Add to trust DB by MD5' logic for reuse
Gentoo's portage also uses MD5s to track which files have changed since
installation.

Added some debug statements for later.
